refactor(shifts): converge ShiftSignups + EventSettings onto ShiftRepository (clear 32 HUM0025 warnings)
VolunteerTrackingRepository read DbContext.ShiftSignups and DbContext.EventSettings
directly, so both tables were shared across two Shifts repositories — violating
HUM0025 ("a table must belong to exactly one repository"). Both repos carried
[Grandfathered("HUM0025", ...)] markers citing
docs/superpowers/specs/2026-05-25-analyzer-consolidation.md, whose intended
resolution is: "route VolunteerTracking through a Shifts read surface."

Executed that convergence:
- Moved GetEligibleBuildSignupsAsync + GetConfirmedShiftsInRangeAsync (both
  fundamentally ShiftSignups -> Shift -> Rota joins filtered by EventSettings)
  from VolunteerTrackingRepository to ShiftRepository, the sole owner of all
  four tables. Declared them (and the EligibleBuildSignup projection) on
  IShiftManagementRepository.
- VolunteerTrackingService already injected IShiftManagementRepository; just
  repointed its 4 calls. VolunteerTrackingExportService swapped its
  IVolunteerTrackingRepository injection for IShiftManagementRepository.
- Removed the 4 now-stale [Grandfathered] attributes and the usings they left
  dead.

ShiftRepository is now the single owner of ShiftSignups + EventSettings, so all
32 HUM0025 warnings for those tables clear (clean build: 127 -> 95 warnings,
0 errors). No schema change. Unit (48) + architecture (229) tests green;
integration-test edits are mechanical DI-resolution swaps (not runnable here —
no Docker — will run in CI).
