+++ src/Humans.Application/Services/Shifts/VolunteerTrackingExportService.cs
@@ -15,19 +15,19 @@ namespace Humans.Application.Services.Shifts;
 /// re-check on <see cref="Humans.Domain.Enums.SignupStatus"/>.
 /// </summary>
 public sealed class VolunteerTrackingExportService(
-    IVolunteerTrackingRepository repository,
+    IShiftManagementRepository shiftManagementRepository,
     IShiftManagementService shiftManagementService,
     IUserServiceRead userService)
     : IVolunteerTrackingExportService, IEarlyEntryProvider
 {
-    private readonly IVolunteerTrackingRepository _repository = repository;
+    private readonly IShiftManagementRepository _shiftManagementRepository = shiftManagementRepository;
     private readonly IShiftManagementService _shiftManagementService = shiftManagementService;
     private readonly IUserServiceRead _userService = userService;
 
     public async Task<VolunteerExportModel> BuildAsync(VolunteerExportRequest request, CancellationToken ct)
     {
         var days = EnumerateDays(request.StartDate, request.EndDate);
-        var shifts = await _repository.GetConfirmedShiftsInRangeAsync(
+        var shifts = await _shiftManagementRepository.GetConfirmedShiftsInRangeAsync(
             request.EventSettingsId, request.StartDate, request.EndDate, request.DepartmentId, ct);
 
         // Resolve team names via the Teams-aware service surface (the Shifts repo
@@ -269,7 +269,7 @@ public sealed class VolunteerTrackingExportService(
 
         var start = es.GateOpeningDate.PlusDays(es.BuildStartOffset);
         var end = es.GateOpeningDate.PlusDays(-1);
-        var rows = await _repository.GetConfirmedShiftsInRangeAsync(es.Id, start, end, departmentId: null, ct);
+        var rows = await _shiftManagementRepository.GetConfirmedShiftsInRangeAsync(es.Id, start, end, departmentId: null, ct);
         if (rows.Count == 0) return [];
 
         var depts = await _shiftManagementService.GetDepartmentsWithRotasAsync(es.Id);
